refactor: relocate `generate_name` to `cw_core` and update imports
