+++ cw_core/lib/generate_name.dart
@@ -1,0 +1,22 @@
+import 'dart:convert';
+import 'dart:math';
+
+import 'package:flutter/services.dart';
+
+extension StringExtension on String {
+  String capitalized() => "${this[0].toUpperCase()}${this.substring(1)}";
+}
+
+Future<String> generateName() async {
+  final randomThing = Random();
+  final adjectiveStringRaw = await rootBundle.loadString("assets/text/Wallet_Adjectives.txt");
+  final nounStringRaw = await rootBundle.loadString("assets/text/Wallet_Nouns.txt");
+
+  final ls = LineSplitter();
+  final adjectives = ls.convert(adjectiveStringRaw);
+  final nouns = ls.convert(nounStringRaw);
+
+  final chosenAdjective = adjectives[randomThing.nextInt(adjectives.length)].capitalized();
+  final chosenNoun = nouns[randomThing.nextInt(nouns.length)].capitalized();
+  return "$chosenAdjective $chosenNoun";
+}
