+++ lib/src/screens/restore/wallet_restore_from_keys_form.dart
@@ -7,7 +7,7 @@ import 'package:cake_wallet/generated/i18n.dart';
 import 'package:cake_wallet/src/widgets/blockchain_height_widget.dart';
 import 'package:cake_wallet/src/widgets/base_text_form_field.dart';
 import 'package:cake_wallet/core/wallet_name_validator.dart';
-import 'package:cake_wallet/entities/generate_name.dart';
+import 'package:cw_core/generate_name.dart';
 import 'package:flutter/services.dart';
 
 class WalletRestoreFromKeysForm extends StatefulWidget {
